Correct processing for begin blocks: - verbatim blocks are catenated together - text blocs are processed for escape sequences
The system remember the type of each block and passes them through the
list of filters. In the end, if the block is a verbatim block or the
last filter is marked as verbatim, the block is enclosed in <pre></pre>.

Text blocks are enclosed in <p></p>.

PPVHF should now support stacking filters.